rename: userStore to UserStore
