+++ frontend/src/Router.jsx
@@ -4,7 +4,7 @@ import { Route, Routes } from 'react-router-dom';
 import { Loading } from '@/components/Common';
 import PublicRoute from './components/Route/PublicRoute';
 import PrivateRoute from './components/Route/PrivateRoute';
-import GoogleAuthCallback from '@/auth/GoogleAuthCallback';
+import GoogleAuthCallback from '@/components/OauthCallback/GoogleAuthCallback';
 
 const Main = lazy(() => import('./pages/Main'));
 const ChannelManager = lazy(() => import('./pages/ChannelManager'));
