+++ frontend/src/components/Route/PrivateRoute/PrivateRoute.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Navigate } from 'react-router-dom';
-import { UserContext } from '@/store/userStore';
+import { UserContext } from '@/store/UserStore';
 
 export default function PrivateRoute({ children }) {
     const {
