+++ frontend/src/components/Route/PublicRoute/PublicRoute.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { ROLE, PAGE_ROLE } from '@/constants/role';
-import { UserContext } from '@/store/userStore';
+import { UserContext } from '@/store/UserStore';
 
 export default function PublicRoute({ component: Component }) {
     const {
