+++ frontend/src/hooks/useSocket.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useContext } from 'react';
 
 import socket from '@/socket';
-import { UserContext } from '@/store/userStore';
+import { UserContext } from '@/store/UserStore';
 import { ModalContext } from '@/store/ModalStore';
 
 import { ChannelEndModalContent } from '@/components/ModalContent';
